+++ app/src/main/java/com/github/autodiag2/elm327emu/sim/ecu/gui.kt
@@ -35,7 +35,7 @@ fun getString(context: Context,resId: Int, vararg formatArgs: Any?): String {
 }
 
 class EcuGui(
-    address: EcuAddress = DEFAULT_ADDRESS.toUByte(),
+    address: EcuAddress = DEFAULT_ADDRESS,
     name: String = EcuType.GUI.toString(),
     private val activity: MainActivity,
 ) : Ecu(EcuType.GUI, address, name, activity) {
