+++ app/src/main/java/com/github/autodiag2/elm327emu/sim/ecu/script.kt
@@ -141,7 +141,7 @@ fun updateScript(script: String, ecu: Ecu) {
     }
 }
 class EcuScript(
-    address: EcuAddress = DEFAULT_ADDRESS.toUByte(),
+    address: EcuAddress = DEFAULT_ADDRESS,
     name: String = EcuType.SCRIPT.toString(),
     private val activity: MainActivity
 ): Ecu(EcuType.SCRIPT, address, name, activity) {
